backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-tenants==3.7.0
django-redis==5.4.0
django-filter==24.3
django-allauth[socialaccount]==65.3.0
dj-rest-auth[with-social]==7.0.1
drf-spectacular==0.27.1
